Minor changes en Factory and implement models and repository for travel and user score
