+++ app/services/user_management_system/__init__.py
@@ -9,7 +9,6 @@ from app.services.user_management_system.contexts.user.infraestructure.router.us
 from app.services.user_management_system.routers.auth import router as authorization_router
 from app.services.user_management_system.routers.automobile import router as automobile_router
 from app.services.user_management_system.routers.drivers import router as driver_router
-from app.services.user_management_system.routers.populate_db import router as db_router
 from app.services.user_management_system.routers.user import router as user_router
 from app.services.user_management_system.shared.authentication.infraestructure.router.authentication_router import (
     router as v2_authorization_router,
@@ -24,7 +23,6 @@ user_management_system_service.include_router(user_router)
 user_management_system_service.include_router(authorization_router)
 user_management_system_service.include_router(driver_router)
 user_management_system_service.include_router(automobile_router)
-user_management_system_service.include_router(db_router)
 user_management_system_service.include_router(v2_user_router)
 user_management_system_service.include_router(v2_authorization_router)
